+++ samples/src/main/java/org/corfudb/samples/BasicCorfuProgram.java
@@ -9,23 +9,19 @@ import org.docopt.Docopt;
 import java.util.Map;
 
 /**
+ * This basic class demonstrates the basics of a Corfu application.
+ *
  * Created by dalia on 12/31/16.
  */
 public abstract class BasicCorfuProgram {
-    private static final String USAGE = "Usage: HelloCorfu [-c <conf>]\n"
-            + "Options:\n"
-            + " -c <conf>     Set the configuration host and port  [default: localhost:9999]\n";
-
     @Setter @Getter
     CorfuRuntime corfuRuntime;
 
     /**
-     * must override this method to initiate any activity
-     */
-    abstract void action();
-
-    /**
+     * Internally, the corfuRuntime interacts with the CorfuDB service over TCP/IP sockets.
+     *
      * @param configurationString specifies the IP:port of the CorfuService
+     *                            The configuration string has format "hostname:port", for example, "localhost:9090".
      * @return a CorfuRuntime object, with which Corfu applications perform all Corfu operations
      */
     private CorfuRuntime getRuntimeAndConnect(String configurationString) {
@@ -33,8 +29,19 @@ public abstract class BasicCorfuProgram {
         return corfuRuntime;
     }
 
+    private static final String USAGE = "Usage: HelloCorfu [-c <conf>]\n"
+            + "Options:\n"
+            + " -c <conf>     Set the configuration host and port  [default: localhost:9999]\n";
+
     /**
-     * boilerplate activity generator
+     * must override this method to initiate any activity
+     */
+    abstract void action();
+
+    /**
+     * boilerplate activity generator, to be invoked from an application's main().
+     *
+     * @param args are the args passed to main
      */
     void start(String[] args) {
         // Parse the options given, using docopt.
@@ -44,16 +51,48 @@ public abstract class BasicCorfuProgram {
                         .parse(args);
         String corfuConfigurationString = (String) opts.get("-c");
 
-
         /**
-         * First, the application needs to instantiate a CorfuRuntime
+         * First, the application needs to instantiate a CorfuRuntime,
+         * which is a Java object that contains all of the Corfu utilities exposed to applications.
+
+         * Above, you will need to point it to a host and port which is running the service.
+         * See {@link https://github.com/CorfuDB/CorfuDB} for instructions on how to deploy Corfu.
          */
         setCorfuRuntime( getRuntimeAndConnect(corfuConfigurationString) );
+
+        /**
+         * Obviously, this application is not doing much yet,
+         * but you can already invoke getRuntimeAndConnect to test if you can connect to a deployed Corfu service.
+         *
+         *
+         * Next, invoke a class-specific activity wrapper named 'action()'.
+         */
         action();
     }
 
     /**
+     * Utility method to instantiate a Corfu object
+     *
+     * A Corfu Stream is a log dedicated specifically to the history of updates of one object.
+     * This method will instantiate a stream by giving it a name,
+     * and then instantiate an object by specifying its class
+     *
+     * @param tClass is the object class
+     * @param name is the name of the stream backing up the object
+     * @param <T> the return class
+     * @return an object instance of type T backed by a stream named 'name'
+     */
+    protected <T> T instantiateCorfuObject(Class<T> tClass, String name) {
+        return getCorfuRuntime().getObjectsView()
+                .build()
+                .setStreamName(name)     // stream name
+                .setType(tClass)        // object class backed by this stream
+                .open();                // instantiate the object!
+    }
+
+    /**
      * Utility method to start a (default type) TX
+     * Can be overriden by classes that require non-default transaction type.
      */
     protected void TXBegin() {
         getCorfuRuntime().getObjectsView().TXBuild()
@@ -65,16 +104,5 @@ public abstract class BasicCorfuProgram {
      */
     protected void TXEnd() {
         getCorfuRuntime().getObjectsView().TXEnd();
-    }
-
-    /**
-     * Utility method to instantiate a Corfu object
-     */
-    protected <T> T instantiateCorfuObject(Class<T> tClass, String name) {
-        return getCorfuRuntime().getObjectsView()
-                .build()
-                .setStreamName(name)     // stream name
-                .setType(tClass)        // object class backed by this stream
-                .open();                // instantiate the object!
     }
 }
