+++ samples/src/main/java/org/corfudb/samples/SimpleAtomicTransaction.java
@@ -81,7 +81,7 @@ public class SimpleAtomicTransaction extends BasicCorfuProgram {
      * main() and standard setup methods are deferred to BasicCorfuProgram
      * @return
      */
-    static BasicCorfuProgram selfFactory() { return new WriteWriteWorkload1(); }
+    static BasicCorfuProgram selfFactory() { return new WriteWriteTXs(); }
     public static void main(String[] args) { selfFactory().start(args); }
 
     /**
