+++ samples/src/main/java/org/corfudb/samples/WriteOnlySimpleTransaction.java
@@ -1,6 +1,5 @@
 package org.corfudb.samples;
 
-import org.corfudb.runtime.CorfuRuntime;
 import org.corfudb.runtime.collections.SMRMap;
 
 import java.util.Map;
@@ -24,7 +23,7 @@ public class WriteOnlySimpleTransaction extends BasicCorfuProgram {
      * main() and standard setup methods are deferred to BasicCorfuProgram
      * @return
      */
-    static BasicCorfuProgram selfFactory() { return new WriteWriteWorkload1(); }
+    static BasicCorfuProgram selfFactory() { return new WriteWriteTXs(); }
     public static void main(String[] args) { selfFactory().start(args); }
